% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixup.R
\docType{package}
\name{mixup}
\alias{mixup}
\alias{mixup-package}
\title{mixup Function}
\usage{
mixup(x1, y1, alpha = 1, concat = FALSE, batch_size = NULL)
}
\description{
mixup Function

mixup: Create Convex Combinations of Pairs of Examples and their Labels for Data Augmentation
}
